src/requirements.txt
#conda activate lbtll-py311

numpy<2.0
torch
torchaudio
pyannote.audio==3.1.1
soundfile
tqdm
openai-whisper